#2939 Meerdere Highlight Boxes: Van patroon naar voorbeeldpagina
