+++ storybook/src/example-pages/Patronen/full-width.content.ts
@@ -1,30 +1,6 @@
 import { DefinitionList, List } from "dso-toolkit";
 import { html, TemplateResult } from "lit-html";
 
-export const mainMenu = [
-  {
-    label: "Home",
-    url: "#",
-    active: true,
-  },
-  {
-    label: "Vergunningscheck",
-    url: "#",
-  },
-  {
-    label: "Aanvragen",
-    url: "#",
-  },
-  {
-    label: "Regels op de kaart",
-    url: "#",
-  },
-  {
-    label: "Mijn Omgevingsloket",
-    url: "#",
-  },
-];
-
 export const definitionList: DefinitionList<TemplateResult> = {
   definitions: [
     {
