+++ storybook/src/example-pages/Patronen/full-width.ts
@@ -1,10 +1,14 @@
 import { html } from "lit-html";
 
 import { examplePageFactory } from "../../example-page-factory";
+
+import { mainMenu } from "../content/main-menu.content";
+
 import { footerPartial } from "../partials/footer";
 import { headerPartial } from "../partials/header";
 import { header } from "../partials/header.content";
-import { definitionList, listGroup, mainMenu } from "./full-width.content";
+
+import { definitionList, listGroup } from "./full-width.content";
 
 examplePageFactory(
   "Patronen",
