+++ storybook/src/example-pages/content/main-menu.content.ts
@@ -1,0 +1,23 @@
+export const mainMenu = [
+  {
+    label: "Home",
+    url: "#",
+    active: true,
+  },
+  {
+    label: "Vergunningscheck",
+    url: "#",
+  },
+  {
+    label: "Aanvragen",
+    url: "#",
+  },
+  {
+    label: "Regels op de kaart",
+    url: "#",
+  },
+  {
+    label: "Mijn Omgevingsloket",
+    url: "#",
+  },
+];
